# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sample/admin-only": {
      "get": {
        "operationId": "SampleController_adminOnly",
        "summary": "Admin-only probe",
        "description": "Returns `{ \"scope\": \"admin\" }` when the JWT carries the `admin` role.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authorized admin caller.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "scope": "admin"
                  },
                  "properties": {
                    "scope": {
                      "type": "string",
                      "example": "admin",
                      "description": "Fixed marker value for this probe endpoint."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the admin role."
          }
        },
        "tags": [
          "sample"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/sample/seeker-only": {
      "get": {
        "operationId": "SampleController_seekerOnly",
        "summary": "Job seeker probe",
        "description": "Returns `{ \"scope\": \"seeker\" }` when the JWT carries the `job_seeker` role.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authorized job seeker caller.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "scope": "seeker"
                  },
                  "properties": {
                    "scope": {
                      "type": "string",
                      "example": "seeker"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the job_seeker role."
          }
        },
        "tags": [
          "sample"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/sample/employer-only": {
      "get": {
        "operationId": "SampleController_employerOnly",
        "summary": "Employer probe",
        "description": "Returns `{ \"scope\": \"employer\" }` when the JWT carries the `employer` role.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authorized employer caller.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "scope": "employer"
                  },
                  "properties": {
                    "scope": {
                      "type": "string",
                      "example": "employer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the employer role."
          }
        },
        "tags": [
          "sample"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Job Portal API",
    "description": "REST API for Job Portal V2: JWT access and refresh tokens, role-based access control, and job postings with pagination and filters.",
    "version": "1.0.0",
    "contact": {
      "name": "Job Portal Team",
      "url": "https://github.com/QCI-Demo-Stage/job-portal-application-v2",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current server (same origin)"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token from login/register (`accessToken` field). Use header `Authorization: Bearer <token>`."
      }
    },
...
  }
}